Support different instalation directory
If the user provides a BIN_DIR variable, it will be used as the
destination of the installed scripts. If nothing is defined, the
previous behavior is used (of installing them on /usr/local/bin).

We do not append that BIN_DIR to the user directory since that
is a user preference best managed by him.